requirements.txt
Pillow>=4.2.0
opencv-python>=4.1.0.25
numpy>=1.16.3